Move Store\UserTrait
