+++ app/Models/Traits/UserStore.php
@@ -3,9 +3,11 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace App\Models\Store;
+namespace App\Models\Traits;
 
-trait UserTrait
+use App\Models\Store\Order;
+
+trait UserStore
 {
     public function orders()
     {
